# refactor: Improve XML performance