--- a/patches/src/main/kotlin/app/revanced/patches/shared/materialyou/BaseMaterialYouPatch.kt
+++ b/patches/src/main/kotlin/app/revanced/patches/shared/materialyou/BaseMaterialYouPatch.kt
@@ -37,8 +37,9 @@ private fun ResourcePatchContext.patchXmlFile(
         val parentList = document.getElementsByTagName(parentNode).item(0) as Element
 
         if (targetNode != null) {
-            for (i in 0 until parentList.childNodes.length) {
-                val node = parentList.childNodes.item(i) as? Element ?: continue
+            val childNodes = parentList.childNodes
+            for (i in 0 until childNodes.length) {
+                val node = childNodes.item(i) as? Element ?: continue
 
                 if (node.nodeName == targetNode && node.hasAttribute(attribute)) {
                     node.getAttributeNode(attribute).textContent = newValue