--- a/patches/src/main/kotlin/app/revanced/patches/shared/translations/BaseTranslationsPatch.kt
+++ b/patches/src/main/kotlin/app/revanced/patches/shared/translations/BaseTranslationsPatch.kt
@@ -123,35 +123,34 @@ fun ResourcePatchContext.baseTranslationsPatch(
     }.toHashSet().toTypedArray()
 
     // Remove unselected app languages from RVX Settings
-    setOf(
-        "revanced_language_entries",
-        "revanced_language_entry_values",
-    ).forEach { attributeName ->
-        document("res/values/arrays.xml").use { document ->
-            with(document) {
-                val nodesToRemove = mutableListOf<Node>()
+    document("res/values/arrays.xml").use { document ->
+        val targetAttributeNames = setOf(
+            "revanced_language_entries",
+            "revanced_language_entry_values",
+        )
+        val nodesToRemove = mutableListOf<Node>()
 
-                val resourcesNode = getElementsByTagName("resources").item(0) as Element
-                for (i in 0 until resourcesNode.childNodes.length) {
-                    val node = resourcesNode.childNodes.item(i) as? Element ?: continue
+        val resourcesNode = document.documentElement
+        val childNodes = resourcesNode.childNodes
+        for (i in 0 until childNodes.length) {
+            val node = childNodes.item(i) as? Element ?: continue
 
-                    if (node.getAttribute("name") == attributeName) {
-                        for (j in 0 until node.childNodes.length) {
-                            val item = node.childNodes.item(j) as? Element ?: continue
-                            val text = item.textContent
-                            val length = text.length
-                            if (!text.endsWith("DEFAULT") && text.subSequence(length - 2, length) !in filteredAppLanguages) {
-                                nodesToRemove.add(item)
-                            }
-                        }
+            if (node.getAttribute("name") in targetAttributeNames) {
+                val itemNodes = node.childNodes
+                for (j in 0 until itemNodes.length) {
+                    val item = itemNodes.item(j) as? Element ?: continue
+                    val text = item.textContent
+                    val length = text.length
+                    if (!text.endsWith("DEFAULT") && text.subSequence(length - 2, length) !in filteredAppLanguages) {
+                        nodesToRemove.add(item)
                     }
                 }
-
-                // Remove the collected nodes (avoids NullPointerException)
-                for (n in nodesToRemove) {
-                    n.parentNode?.removeChild(n)
-                }
             }
+        }
+
+        // Remove the collected nodes (avoids NullPointerException)
+        for (n in nodesToRemove) {
+            n.parentNode?.removeChild(n)
         }
     }
 }