--- a/patches/src/main/kotlin/app/revanced/patches/youtube/general/snackbar/SnackBarComponentsPatch.kt
+++ b/patches/src/main/kotlin/app/revanced/patches/youtube/general/snackbar/SnackBarComponentsPatch.kt
@@ -348,14 +348,16 @@ val snackBarComponentsPatch = resourcePatch(
         }
 
         document("res/values/dimens.xml").use { document ->
-            val resourcesNode = document.getElementsByTagName("resources").item(0) as Element
+            val resourcesNode = document.documentElement
+            val childNodes = resourcesNode.childNodes
 
-            for (i in 0 until resourcesNode.childNodes.length) {
-                val node = resourcesNode.childNodes.item(i) as? Element ?: continue
+            for (i in 0 until childNodes.length) {
+                val node = childNodes.item(i) as? Element ?: continue
                 val dimenName = node.getAttribute("name")
 
                 if (dimenName.equals("snackbar_corner_radius")) {
                     node.textContent = cornerRadius
+                    break
                 }
             }
         }