--- a/patches/src/main/kotlin/app/revanced/patches/youtube/layout/theme/ThemePatch.kt
+++ b/patches/src/main/kotlin/app/revanced/patches/youtube/layout/theme/ThemePatch.kt
@@ -77,10 +77,11 @@ val themePatch = resourcePatch(
 
         arrayOf("values", "values-v31").forEach { path ->
             document("res/$path/colors.xml").use { document ->
-                val resourcesNode = document.getElementsByTagName("resources").item(0) as Element
+                val resourcesNode = document.documentElement
+                val childNodes = resourcesNode.childNodes
 
-                for (i in 0 until resourcesNode.childNodes.length) {
-                    val node = resourcesNode.childNodes.item(i) as? Element ?: continue
+                for (i in 0 until childNodes.length) {
+                    val node = childNodes.item(i) as? Element ?: continue
 
                     node.textContent = when (node.getAttribute("name")) {
                         "yt_black0", "yt_black1", "yt_black1_opacity95", "yt_black1_opacity98", "yt_black2", "yt_black3",