Improve the performance of XML operations by not calling `node.getChildNodes()` multiple times in a loop.

The result of `node.getChildNodes()` should be stored in a variable before starting a for loop.

I compared the performance before PR and after PR, using a Snapdragon 662+4GB RAM phone and a Snapdragon 410+2GB RAM phone
This resulted in significant improvements for some XML files, especially those with many child nodes.

```
https://github.com/inotia00/revanced-patches/blob/b1776cf00729216c001a8dc6720468c08666936d/patches/src/main/kotlin/app/revanced/patches/youtube/layout/theme/ThemePatch.kt#L79-L90
[res/values/colors.xml]
SD662: 637 ms -> 3 ms
SD410: 2971 ms -> 6 ms

https://github.com/inotia00/revanced-patches/blob/b1776cf00729216c001a8dc6720468c08666936d/patches/src/main/kotlin/app/revanced/patches/youtube/utils/settings/SettingsPatch.kt#L227-L244
[res/values/styles.xml]
SD662: 1107 ms -> 10 ms
SD410: 6401 ms -> 35 ms

https://github.com/inotia00/revanced-patches/blob/b1776cf00729216c001a8dc6720468c08666936d/patches/src/main/kotlin/app/revanced/patches/youtube/general/snackbar/SnackBarComponentsPatch.kt#L350-L361
[res/values/dimens.xml]
SD662: 4837 ms -> 11 ms
SD410: 24929 ms -> 13 ms
```

In addition, I did some refactoring:

- Usually, to get the root node of a document, use `document.getDocumentElement()`
- Avoid iterating one XML twice
- break a for loop if possible

I tested this change with 19.44.36, 19.16.36, and 18.29.36 and confirmed that this changes don't change the result resources.
I compared the `patched-temporary-files` dir of current patches and my PR branch patches using WinMerge and those were identical.