@inotia00 2025-02-11
This PR includes a small refactoring for ResourcePatch.

Can this PR includes a class `java.nio.file.Files` check for Android 5 - 7?