@kitadai31 2025-02-11
No, I'm just going to push a second PR that is for A5-7 support